refactor: RedshiftStatementHook -> RedshiftSQLHook RedshiftOperator -> RedshiftSQLOperator
