+++ airflow/providers/amazon/aws/example_dags/example_redshift.py
@@ -16,17 +16,17 @@
 # specific language governing permissions and limitations
 # under the License.
 """
-This is an example dag for using `RedshiftOperator` to authenticate with Amazon Redshift
+This is an example dag for using `RedshiftSQLOperator` to authenticate with Amazon Redshift
 then execute a simple select statement
 """
 # [START redshift_operator_howto_guide]
 from airflow import DAG
-from airflow.providers.amazon.aws.operators.redshift import RedshiftOperator
+from airflow.providers.amazon.aws.operators.redshift import RedshiftSQLOperator
 from airflow.utils.dates import days_ago
 
 with DAG(dag_id="example_redshift", start_date=days_ago(1), schedule_interval=None, tags=['example']) as dag:
-    # [START howto_operator_s3_to_redshift_create_table]
-    setup__task_create_table = RedshiftOperator(
+    # [START howto_operator_redshift_create_table]
+    setup__task_create_table = RedshiftSQLOperator(
         task_id='setup__create_table',
         sql="""
             CREATE TABLE IF NOT EXISTS fruit (
@@ -36,9 +36,9 @@ with DAG(dag_id="example_redshift", start_date=days_ago(1), schedule_interval=No
             );
         """,
     )
-    # [END howto_operator_s3_to_redshift_create_table]
-    # [START howto_operator_s3_to_redshift_populate_table]
-    task_insert_data = RedshiftOperator(
+    # [END howto_operator_redshift_create_table]
+    # [START howto_operator_redshift_populate_table]
+    task_insert_data = RedshiftSQLOperator(
         task_id='task_insert_data',
         sql=[
             "INSERT INTO fruit VALUES ( 1, 'Banana', 'Yellow');",
@@ -49,17 +49,19 @@ with DAG(dag_id="example_redshift", start_date=days_ago(1), schedule_interval=No
             "INSERT INTO fruit VALUES ( 6, 'Strawberry', 'Red');",
         ],
     )
-    # [END howto_operator_s3_to_redshift_populate_table]
-    # [START howto_operator_s3_to_redshift_get_all_rows]
-    task_get_all_table_data = RedshiftOperator(task_id='task_get_all_table_data', sql="SELECT * FROM fruit;")
-    # [END howto_operator_s3_to_redshift_get_all_rows]
-    # [START howto_operator_s3_to_redshift_get_with_filter]
-    task_get_with_filter = RedshiftOperator(
+    # [END howto_operator_redshift_populate_table]
+    # [START howto_operator_redshift_get_all_rows]
+    task_get_all_table_data = RedshiftSQLOperator(
+        task_id='task_get_all_table_data', sql="SELECT * FROM fruit;"
+    )
+    # [END howto_operator_redshift_get_all_rows]
+    # [START howto_operator_redshift_get_with_filter]
+    task_get_with_filter = RedshiftSQLOperator(
         task_id='task_get_with_filter',
         sql="SELECT * FROM fruit WHERE color = '{{ params.color }}';",
         params={'color': 'Red'},
     )
-    # [END howto_operator_s3_to_redshift_get_with_filter]
+    # [END howto_operator_redshift_get_with_filter]
 
     setup__task_create_table >> task_insert_data >> task_get_all_table_data >> task_get_with_filter
 # [END redshift_operator_howto_guide]
